+++ test/strategy/hungry_test.exs
@@ -6,18 +6,32 @@ defmodule Test.Strategy.Hungry do
 
   alias __MODULE__.Adder, as: HA
 
-  test "basic operation" do
+  setup do
     HA.initialize()
+    :ok
+  end
+
+  test "basic operation" do
     assert HA.consume([1,2,"cat"]) == [ 3, 6, "catcat" ]
   end
 
   test "supports into" do
-    HA.initialize()
-    assert HA.consume(a: :b, c: :d) == [ b: :a, d: :c ]
+    assert HA.consume([ a: :b, c: :d ], into: %{}) == %{ b: :a, d: :c }
+  end
+
+  test "into function" do
+    HA.consume([:a, :b, :c, :d, :e, :f], into: fn x ->
+      assert x in ~w{ a b c d e f }
+    end)
+  end
+
+  test "into stream" do
+    stream = HA.consume([:a, :b, :c, :d, :e, :f], into: :stream)
+    assert %Stream{} = stream
+    assert Enum.into(stream, []) == ~w{ a b c d e f }
   end
 
   test "schedules workers" do
-    HA.initialize()
     assert HA.consume(1..100) == (1..100 |> Enum.map(&(&1*3)))
   end
 
@@ -26,7 +40,7 @@ defmodule Test.Strategy.Hungry do
 
     use Component.Strategy.Hungry,
         show_code:           false,
-        default_concurrency: 5
+        concurrency: 5
 
     def process(val) when is_number(val) do
       val * 3
@@ -34,6 +48,10 @@ defmodule Test.Strategy.Hungry do
 
     def process(val) when is_binary(val) do
       val <> val
+    end
+
+    def process(val) when is_atom(val) do
+      "#{val}"
     end
 
     def process({a, b}), do: { b, a }